Refactor cache folder creation to its own function
Move the code from `pooch.create` to `pooch.utils.make_local_storage`.
The function joins the path, adds the version string, checks the
environment variable, expands the user path, checks if the path is
writable (now printing different messages if we can't create the folder
vs can't create a file), and converts the string path to `pathlib.Path`.
This will make it easier to reuse this part of the code is needed.
